[Gold V] Title: 다각형의 면적, Time: 8 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Gold/2166. 다각형의 면적/다각형의 면적.cc
@@ -16,7 +16,7 @@ long long cross_product(point origin, point p, point q) {
 
 int main() {
   int N;
-  long double area = 0;
+  double area = 0;
 
   cin >> N;
